+++ src/com/intellij/idea/plugin/hybris/impex/lang/folding/ImpexFoldingBuilder.java
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,13 +18,10 @@
 
 package com.intellij.idea.plugin.hybris.impex.lang.folding;
 
-import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent;
 import com.intellij.lang.ASTNode;
-import com.intellij.lang.folding.FoldingBuilderEx;
 import com.intellij.lang.folding.FoldingDescriptor;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.FoldingGroup;
-import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.SyntaxTraverser;
 import org.apache.commons.lang3.Validate;
@@ -36,26 +33,17 @@ import java.util.*;
 
 import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.isLineBreak;
 
-public class ImpexFoldingBuilder extends FoldingBuilderEx {
+public class ImpexFoldingBuilder extends AbstractImpExFoldingBuilder {
 
     private static final String GROUP_NAME = "ImpEx";
 
-    private static final FoldingDescriptor[] EMPTY_ARRAY = new FoldingDescriptor[0];
-
     @NotNull
     @Override
-    public FoldingDescriptor[] buildFoldRegions(
+    public FoldingDescriptor[] buildFoldRegionsInternal(
         @NotNull final PsiElement root,
         @NotNull final Document document,
         final boolean quick
     ) {
-        if (this.isFoldingDisabled(root.getProject())) {
-            return EMPTY_ARRAY;
-        }
-
-        Validate.notNull(root);
-        Validate.notNull(document);
-
         final Collection<PsiElement> psiElements = this.findFoldingBlocksAndLineBreaks(root);
 
         FoldingGroup currentLineGroup = FoldingGroup.newGroup(GROUP_NAME);
@@ -78,15 +66,6 @@ public class ImpexFoldingBuilder extends FoldingBuilderEx {
         }
 
         return descriptors.toArray(new FoldingDescriptor[descriptors.size()]);
-    }
-
-
-    @Contract(pure = true)
-    protected boolean isFoldingDisabled(final @NotNull Project project) {
-        return !HybrisProjectSettingsComponent.getInstance(project).getState()
-            .getImpexSettings()
-            .getFolding()
-            .getEnabled();
     }
 
     @NotNull