+++ src/com/intellij/idea/plugin/hybris/impex/lang/folding/ImpexFoldingLinesBuilder.java
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,14 +20,10 @@ package com.intellij.idea.plugin.hybris.impex.lang.folding;
 
 import com.intellij.idea.plugin.hybris.impex.lang.folding.smart.ImpexFoldingLinesFilter;
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes;
-import com.intellij.idea.plugin.hybris.settings.HybrisApplicationSettingsComponent;
-import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent;
 import com.intellij.lang.ASTNode;
-import com.intellij.lang.folding.FoldingBuilderEx;
 import com.intellij.lang.folding.FoldingDescriptor;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.FoldingGroup;
-import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.SyntaxTraverser;
 import org.apache.commons.lang3.Validate;
@@ -40,33 +36,19 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.getPrevNonWhitespaceElement;
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.isHeaderLine;
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.isImpexValueLine;
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.isUserRightsMacros;
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.nextElementIsHeaderLine;
-import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.nextElementIsUserRightsMacros;
+import static com.intellij.idea.plugin.hybris.impex.utils.ImpexPsiUtils.*;
 
-public class ImpexFoldingLinesBuilder extends FoldingBuilderEx {
+public class ImpexFoldingLinesBuilder extends AbstractImpExFoldingBuilder {
 
     private static final String LINE_GROUP_NAME = "impex_fold_line";
 
-    private static final FoldingDescriptor[] EMPTY_ARRAY = new FoldingDescriptor[0];
-
     @NotNull
     @Override
-    public FoldingDescriptor[] buildFoldRegions(
+    public FoldingDescriptor[] buildFoldRegionsInternal(
         @NotNull final PsiElement root,
         @NotNull final Document document,
         final boolean quick
     ) {
-        if (this.isFoldingDisabled(root.getProject())) {
-            return EMPTY_ARRAY;
-        }
-
-        Validate.notNull(root);
-        Validate.notNull(document);
-
         FoldingGroup currentLineGroup = FoldingGroup.newGroup(LINE_GROUP_NAME);
 
         final List<PsiElement> foldingBlocks = foldingLines(root);
@@ -117,15 +99,6 @@ public class ImpexFoldingLinesBuilder extends FoldingBuilderEx {
         }
 
         return descriptors.toArray(new FoldingDescriptor[descriptors.size()]);
-    }
-
-
-    @Contract(pure = true)
-    protected boolean isFoldingDisabled(final @NotNull Project project) {
-        return !HybrisProjectSettingsComponent.getInstance(project).getState()
-            .getImpexSettings()
-            .getFolding()
-            .getEnabled();
     }
 
     @NotNull