+++ src/com/intellij/idea/plugin/hybris/system/type/model/CustomProperty.java
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -38,6 +38,7 @@ import org.jetbrains.annotations.NotNull;
 public interface CustomProperty extends DomElement {
 
     String NAME = "name";
+    String VALUE = "value";
 
     /**
      * Returns the value of the name child.
@@ -64,7 +65,7 @@ public interface CustomProperty extends DomElement {
      * @return the value of the value child.
      */
     @NotNull
-    @SubTag("value")
+    @SubTag(VALUE)
     @Required
     GenericDomValue<String> getValue();